requirements.txt
# Morning Briefing Skill Dependencies

# Core
requests>=2.31.0
PyYAML>=6.0.1

# RSS/Feed parsing
feedparser>=6.0.10

# PDF generation
reportlab>=4.0.0

# Markdown to HTML (for email distribution)
markdown>=3.5.0

# Machine learning (for paper scoring)
scikit-learn>=1.3.0

# HTML sanitization (for email distribution)
nh3>=0.2.14

# XML security (optional, falls back to stdlib)
defusedxml>=0.7.1

# OpenTelemetry tracing
opentelemetry-api>=1.25.0
opentelemetry-sdk>=1.25.0
opentelemetry-exporter-otlp-proto-http>=1.25.0
opentelemetry-instrumentation-requests>=0.46b0

# NotebookLM podcast generation
notebooklm-py>=0.3.0

